requirements.txt
boto3==1.34.0
discord.py==2.3.2
python-dotenv==1.0.0
